requirements.txt
pika==1.3.1
fastapi==0.86.0
aio-pika==8.2.4
uvicorn==0.20
asyncio==3.4.3
SQLAlchemy==1.4.42
pydantic==1.10.2
requests==2.28.1
pytest==7.1.3
mock==4.0.3
aiosqlite==0.17.0
timing-asgi==0.2.1
aiohttp==3.8.3
pytest-asyncio==0.20.1
asyncpg==0.27.0
email-validator==1.3.0
scout-apm==2.26.1
alembic==1.8.1
python-dotenv~=0.21.0
names==0.3.0
namegenerator==1.0.6
fastapi-pagination==0.11.0
sqlakeyset==1.0.1659142803
locust==2.14.0
aioredis==2.0.1
hiredis==2.1.0
timedelta==2020.12.3
pylint
